refactor for Component folder
