+++ src/components/Job-Table/Job-Table.js
@@ -4,10 +4,10 @@ import './Job-Table.css';
 import React, { Component, PropTypes } from 'react';
 import classnames from 'classnames';
 
-import Table from '../A10UI/Table/Table';
-import Td from '../A10UI/Table/Td';
-import EditButton from '../A10UI/Button/Edit-Button';
-import DeleteButton from '../A10UI/Button/Delete-Button';
+import Table from '../A10-UI/Table/Table';
+import Td from '../A10-UI/Table/Td';
+import EditButton from '../A10-UI/Button/Edit-Button';
+import DeleteButton from '../A10-UI/Button/Delete-Button';
 
 let TableHeaders = ({ titleKeyMap, onSortHandler, sortBy, enableSort }) => {
 	let headerHtml = titleKeyMap.map((headerObj, index) => {
